requirements.txt
streamlit>=1.32.0
tensorflow>=2.15.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0
crewai>=0.28.0
groq>=0.5.0
scikit-learn>=1.4.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.2.0
pillow>=10.2.0
requests>=2.31.0
duckduckgo_search>=5.3.0
ultralytics>=8.1.0
pydantic>=2.6.0
nest_asyncio>=1.6.0
beautifulsoup4>=4.12.0
